+++ src/eslint-defaults.js
@@ -4,6 +4,6 @@ const path = require('path');
 
 module.exports = {
   configFile: path.join(__dirname, '../.eslintrc'),
-  extensions: ['.js', '.jsx'],
+  extensions: ['.js'],
   files: require('./lib/global-source-files')
 };